+++ xarray/core/dataset.py
@@ -3134,7 +3134,7 @@ class Dataset(DataWithCoords, DatasetArithmetic, Mapping):
 
     def _rename_indexes(self, name_dict, dims_set):
         if self._indexes is None:
-            return None
+            return {}
         indexes = {}
         for k, v in self.indexes.items():
             new_name = name_dict.get(k, k)
@@ -3152,6 +3152,24 @@ class Dataset(DataWithCoords, DatasetArithmetic, Mapping):
         variables, coord_names = self._rename_vars(name_dict, dims_dict)
         dims = self._rename_dims(dims_dict)
         indexes = self._rename_indexes(name_dict, dims.keys())
+
+        # Variable could be renamed to an existing dimension name
+        # in this case, convert to IndexVariable and set indexes
+        # GH4107
+        for name in set(self.dims) & set(variables) - set(indexes):
+            variables[name] = variables[name].to_index_variable()
+            indexes[name] = variables[name].to_index()
+            coord_names.add(name)
+
+        # rename_dims was called to rename an indexed dimension
+        # the new renamed dimension is unindexed
+        # remove the old variable from indexes and convert to a base variable
+        if indexes:
+            for name, newname in dims_dict.items():
+                if name != newname and name in indexes:
+                    variables[name] = variables[name].to_base_variable()
+                    del indexes[name]
+
         return variables, coord_names, dims, indexes
 
     def rename(